Rename ndmsmt submodules
