+++ src/accumulators/ndm_smt/ndm_smt_config.rs
